EMP-553: Interesting refactor moving query filter criterias inside repository package since they are port of the query construction
EMP-553: Interesting refactor moving query filter criteria inside repository package since they are port of the query construction instead of configuration
