--- equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/config/CrmFormModelInterface.java
@@ -1,3 +1,0 @@
-package uk.gov.justice.laa.crime.equinity.historicaldata.config;
-
-public interface CrmFormModelInterface {}
