+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/mapper/CrmFormsViewMapper.java
@@ -3,7 +3,7 @@ package uk.gov.justice.laa.crime.equinity.historicaldata.mapper;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.springframework.data.domain.Page;
-import uk.gov.justice.laa.crime.equinity.historicaldata.config.CrmFormModelInterface;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.data.CrmFormDataModelInterface;
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.SearchCrmFormDTO;
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.SearchResultDTO;
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.SearchPagingDTO;
@@ -15,7 +15,7 @@ import java.util.List;
 public interface CrmFormsViewMapper {
     @Mapping(target="results", source="crmFormsViewModels")
     @Mapping(target="paging", source="page")
-    SearchResultDTO getDTOsFromModel(List<CrmFormViewModel> crmFormsViewModels, Page<CrmFormModelInterface> page);
+    SearchResultDTO getDTOsFromModel(List<CrmFormViewModel> crmFormsViewModels, Page<CrmFormDataModelInterface> page);
 
     @Mapping(target="usn", source="USN")
     @Mapping(target="type", source="type")
@@ -32,6 +32,6 @@ public interface CrmFormsViewMapper {
     @Mapping(target="total", source="page.totalPages")
     @Mapping(target="itemsPage", source="page.numberOfElements")
     @Mapping(target="itemsTotal", source="page.totalElements")
-    SearchPagingDTO getPagingDTOFromPage(Page<CrmFormModelInterface> page);
+    SearchPagingDTO getPagingDTOFromPage(Page<CrmFormDataModelInterface> page);
 }
