+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/model/data/CrmFormDataModelInterface.java
@@ -1,0 +1,3 @@
+package uk.gov.justice.laa.crime.equinity.historicaldata.model.data;
+
+public interface CrmFormDataModelInterface {}
