+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/model/data/CrmFormViewModel.java
@@ -5,14 +5,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.Immutable;
-import uk.gov.justice.laa.crime.equinity.historicaldata.config.CrmFormModelInterface;
 
 @Immutable @Entity
 @Table(name = "CrmFormsView")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class CrmFormViewModel implements CrmFormModelInterface {
+public class CrmFormViewModel implements CrmFormDataModelInterface {
 
     @Id
     @Column(name="USN")
