+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/repository/CrmFormsViewRepository.java
@@ -2,9 +2,9 @@ package uk.gov.justice.laa.crime.equinity.historicaldata.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import uk.gov.justice.laa.crime.equinity.historicaldata.config.CrmFormModelInterface;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.data.CrmFormDataModelInterface;
 import uk.gov.justice.laa.crime.equinity.historicaldata.model.data.CrmFormViewModel;
 
 
-public interface CrmFormsViewRepository extends JpaRepository<CrmFormViewModel, Integer>, JpaSpecificationExecutor<CrmFormModelInterface> {
+public interface CrmFormsViewRepository extends JpaRepository<CrmFormViewModel, Integer>, JpaSpecificationExecutor<CrmFormDataModelInterface> {
 }
