+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/repository/criteria/input/Crm14CaseSummaryReportCriteriaDTO.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.equinity.historicaldata.config;
+package uk.gov.justice.laa.crime.equinity.historicaldata.repository.criteria.input;
 
 import jakarta.annotation.Nullable;
 import jakarta.validation.constraints.NotNull;
